move: Upload.tsx를 pages > Pra → pages로 이동
